Resolve "Store configuration pages"

Closes #16

See merge request edgsolutions-engineering/clear-fusion!48

+++ HumanitarianAssistance.Application/Store/Models/PurchasedGeneratorModel.cs
@@ -3,11 +3,11 @@ namespace HumanitarianAssistance.Application.Store.Models
     public class PurchasedGeneratorModel
     {
         public long? Id { get; set; }
-        public int Voltage { get; set; }
-        public int StartingUsage { get; set; }
-        public int IncurredUsage { get; set; }
-        public int FuelConsumptionRate { get; set; }
-        public int MobilOilConsumptionRate { get; set; }
+        public double Voltage { get; set; }
+        public double StartingUsage { get; set; }
+        public double IncurredUsage { get; set; }
+        public double FuelConsumptionRate { get; set; }
+        public double MobilOilConsumptionRate { get; set; }
         public int OfficeId { get; set; }
         public int ModelYear { get; set; }
         public long? PurchaseId { get; set; }
